+++ src/arrayIO/hdf5.h
@@ -17,7 +17,7 @@ template <class T> class H5NumWriter : public BulkNumWriter<T> {
     HighFive::DataType datatype = HighFive::create_datatype<T>();
 
     static HighFive::DataSet
-    createH5DataSet(HighFive::Group group, std::string group_path, uint32_t chunk_size) {
+    createH5DataSet(HighFive::Group group, std::string group_path, uint64_t chunk_size) {
         HighFive::SilenceHDF5 s;
         // Create a dataspace with initial shape and max shape
         HighFive::DataSpace dataspace({0}, {HighFive::DataSpace::UNLIMITED});
@@ -41,11 +41,11 @@ template <class T> class H5NumWriter : public BulkNumWriter<T> {
     }
 
   public:
-    H5NumWriter(const HighFive::Group &group, std::string path, uint32_t chunk_size = 1024)
+    H5NumWriter(const HighFive::Group &group, std::string path, uint64_t chunk_size = 1024)
         : dataset(createH5DataSet(group, path, chunk_size)) {}
 
-    uint32_t write(T *in, uint32_t count) override {
-        uint32_t cur_size = dataset.getDimensions()[0];
+    uint64_t write(T *in, uint64_t count) override {
+        uint64_t cur_size = dataset.getDimensions()[0];
         dataset.resize({cur_size + count});
         dataset.select({cur_size}, {count}).write_raw(in, datatype);
         return count;
@@ -64,14 +64,14 @@ template <class T> class H5NumReader : public BulkNumReader<T> {
     H5NumReader(const HighFive::Group &group, std::string path) : dataset(group.getDataSet(path)) {}
 
     // Return total number of integers in the reader
-    uint32_t size() const override { return dataset.getDimensions()[0]; }
+    uint64_t size() const override { return dataset.getDimensions()[0]; }
 
     // Change the next load to start at index pos
-    void seek(uint32_t new_pos) override { pos = new_pos; }
+    void seek(uint64_t new_pos) override { pos = new_pos; }
 
     // Copy up to `count` integers into `out`, returning the actual number copied.
     // Will always load >0 unless there is no more input
-    uint32_t load(T *out, uint32_t count) override {
+    uint64_t load(T *out, uint64_t count) override {
         dataset.select({pos}, {count}).read(out, datatype);
         pos += count;
         return count;
@@ -86,8 +86,8 @@ class H5StringReader : public StringReader {
 
   public:
     H5StringReader(const HighFive::Group &group, std::string path);
-    const char *get(uint32_t idx) const override;
-    uint32_t size() const override;
+    const char *get(uint64_t idx) const override;
+    uint64_t size() const override;
 };
 
 class H5StringWriter : public StringWriter {
@@ -103,12 +103,12 @@ class H5StringWriter : public StringWriter {
 class H5WriterBuilder final : public WriterBuilder {
   protected:
     HighFive::Group group;
-    uint32_t buffer_size;
-    uint32_t chunk_size;
+    uint64_t buffer_size;
+    uint64_t chunk_size;
 
   public:
     H5WriterBuilder(
-        std::string file, std::string group, uint32_t buffer_size = 8192, uint32_t chunk_size = 1024, bool allow_exists = false
+        std::string file, std::string group, uint64_t buffer_size = 8192, uint64_t chunk_size = 1024, bool allow_exists = false
     );
     UIntWriter createUIntWriter(std::string name) override;
     ULongWriter createULongWriter(std::string name) override;
@@ -121,12 +121,12 @@ class H5WriterBuilder final : public WriterBuilder {
 
 class H5ReaderBuilder final : public ReaderBuilder {
     HighFive::Group group;
-    uint32_t buffer_size;
-    uint32_t read_size;
+    uint64_t buffer_size;
+    uint64_t read_size;
 
   public:
     H5ReaderBuilder(
-        std::string file, std::string group, uint32_t buffer_size, uint32_t read_size = 1024
+        std::string file, std::string group, uint64_t buffer_size, uint64_t read_size = 1024
     );
     UIntReader openUIntReader(std::string name) override;
     ULongReader openULongReader(std::string name) override;
